\name{LOG}
\alias{LOG}
\alias{log}
\alias{DDMarker}
\title{
LOG
}
\description{
The logs of Diagnose and Detect Markers in Extracellular Circulating \cr \cr
logs(); \cr \cr
}
\details{
logs();
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \email{yujoan_2001@163.com} \cr 
Huansheng Cao (UGA) \email{hshcao@uga.edu} \cr 
Guoqing Liu (IMUST & UGA) \email{gqliu@uga.edu} \email{gqliu1010@163.com} \cr 
Xiufeng Liu (GZUCM & UGA) \email{xfliu@uga.edu} \email{liu_xf@gzucm.edu.cn} \cr 
Hao Wu (BIT & UGA) \email{wuhao@uga.edu} \email{wuhao@bit.edu.cn} \cr 
Yan Wang (JLU & UGA) \email{wy6868@hotmail.com} \cr 
Ying Xu (JLU & UGA) \email{xyn@uga.edu} \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
citation("DDMarkerFAST");
}
\keyword{ DDMarkerFAST }
\keyword{ DDMarkerP }
\keyword{ LOG }
\seealso{
\code{\link{DDMarkerFAST-package}} \code{\link{DDMarkerFAST-method}} \code{\link{DDMarkerP-method}} \code{\link{NAR}} \cr
}
\examples{

logs();
# print the logs of DDMarker

## LOG
# DDMarker 1.0, 05/27/2015, Recomputed the blood and urine predicted biomarkers from the latest datebase;
# DDMarker 1.1, 08/27/2015, DDMarker can entails the sequence of proteins;
# DDMarker 1.2, 12/28/2015, adds a visualization function of biomarkers; isoform proteins can be diagnosed; starts the prediction of extracellular circulating in a large area;
# DDMarker 1.3, 01/12/2016, more illustrates are upload as the draft;
# DDMarker 1.4, 03/12/2016, adds the medical guidance in the results;
# DDMarker 1.4.1, 04/01/2016, bug fixed;
# DDMarker 1.4.2, 05/02/2016, bug fixed;
# DDMarker 1.4.3, 05/23/2016, bug fixed;
# DDMarker 1.5, 05/27/2016, bug fixed;
# DDMarker 2.0, 07/07/2016, DDMarker2 published;
# DDMarker 2.1, 07/12/2016, adds 4 methods in DDMarker FAST method.

}
