\name{NAR}
\alias{NAR}
\alias{nar}
\title{
NAR
}
\description{
The famous features, which were used in nucleic acids research and centers for disease control and prevention \cr \cr
nar();
}
\details{
nar();
}
\author{
Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
Qiong Yu (JLU & UGA) \email{yuqiong@uga.edu} \email{yujoan_2001@163.com} \cr 
Huansheng Cao (UGA) \email{hshcao@uga.edu} \cr 
Guoqing Liu (IMUST & UGA) \email{gqliu@uga.edu} \email{gqliu1010@163.com} \cr 
Xiufeng Liu (GZUCM & UGA) \email{xfliu@uga.edu} \email{liu_xf@gzucm.edu.cn} \cr 
Hao Wu (BIT & UGA) \email{wuhao@uga.edu} \email{wuhao@bit.edu.cn} \cr 
Yan Wang (JLU & UGA) \email{wy6868@hotmail.com} \cr 
Ying Xu (JLU & UGA) \email{xyn@uga.edu} \email{xyn@bmb.uga.edu} \cr \cr 
Maintainer: Yu Shang (JLU & UGA) \email{yushang@uga.edu} \cr 
}
\references{
citation("DDMarkerFAST");
\cr 
\code{[1]} Juan Cui, et al. (2011) \emph{An integrated transcriptomic and computational analysis for biomarker identification in gastric cancer} Nucleic Acids Research, 39: 1197-1207 \cr
\code{[2]} Juan Cui, et al. (2008) \emph{Computational prediction of human proteins that can be secreted into the bloodstream} BIOINFORMATICS, Vol.24 no. 20 2008 pages 2370-2375 \cr
\code{[3]} \url{http://bioinfosrv1.bmb.uga.edu/DMarker/} \cr
}
\keyword{ DDMarkerFAST }
\keyword{ DDMarkerP }
\keyword{ NAR }
\seealso{
\code{\link{DDMarkerFAST-package}} \code{\link{DDMarkerFAST-method}} \code{\link{DDMarkerP-method}} \code{\link{DDMarker}} \cr
}
\examples{

nar();
# print the famous features, which were used in Nucleic Acids Research and Centers for Disease Control and Prevention
# See the references

## Features used in DDMarker:
## 
# http://bidd.cz3.nus.edu.sg/cgi-bin/prof/protein/profnew.cgi
# http://www.expasy.org/proteomics/protein_structure
# http://molbiol-tools.ca/Protein_secondary_structure.htm
# http://www.cbs.dtu.dk/services/
# http://coot.embl.de/cgi/sscp_serv.pl
# http://phobius.sbc.su.se/cgi-bin/predict.pl
# http://bip.weizmann.ac.il/fldbin/findex
# http://www.cbs.dtu.dk/services/TatP/
# http://bmbpcu36.leeds.ac.uk/~andy/betaBarrel/AACompPred/aaTMB_Hunt.cgi
# http://jing.cz3.nus.edu.sg/cgi-bin/prof/prof.cgi
# http://www.cbs.dtu.dk/services/NetNGlyc/
# http://www.cbs.dtu.dk/services/NetOGlyc/
# http://www.scfbio-iitd.res.in/software/proteomics/rg.jsp
# http://web.expasy.org/cgi-bin/compute_pi/pi_tool

}
